tools/blister-config-extractor/elastic/blister_config_extractor/requirements.txt
malduck
yara-python
pefile
